Split tests up into multiple modules
